Extract 'tools' graphql to /graphql and /hooks folders, remove top-level graphql folder
