+++ src/features/tools/campaign-codes/graphql/add-partner-percentage-discount-code.graphql
