+++ src/features/tools/campaign-codes/graphql/add-partner-visible-no-discount-code.graphql
