+++ src/features/tools/campaign-codes/graphql/available-marketing-channels.graphql
