+++ src/features/tools/campaign-codes/graphql/create-campaign-partner.graphql
