+++ src/features/tools/campaign-codes/graphql/get-partner-campaigns.graphql
