+++ src/features/tools/campaign-codes/graphql/set-campaign-marketing-channel.graphql
