+++ src/features/tools/employees/graphql/remove-employee.graphql
