--- src/graphql/tools/unsign-member.graphql
@@ -1,3 +1,0 @@
-mutation UnsignMember($ssn: String!) {
-  unsignMember(ssn: $ssn)
-}
